Add Sounds Menu
Lets the user choose one of four unique sounds to represent their "buzz in" sound (plus a "no sound" option). This will help distinguish who buzzed in first, if multiple users are playing in the same space.

There is currently an issue where the buzz-in sound plays after a decent delay, that didn't happen when there was only one sound. Will need to troubleshoot that.